requirements.txt
Pillow
requests
moviepy
selenium
webdriver-manager
bs4
google-api-python-client
google-auth
google-auth-oauthlib